requirements.txt
Django>=4.2,<6.0
reportlab>=4.0
gunicorn>=21.2
whitenoise>=6.5
dj-database-url>=2.1
psycopg2-binary>=2.9
python-dotenv>=1.0
